Renamed DataAggregator to StatisticsAggregator
